pc_controller/requirements.txt
flask>=3.0
websocket-client>=1.7
requests>=2.31
opencv-python>=4.8
numpy>=1.24
torch>=2.1
torchvision>=0.16
Pillow>=10.0
